Refactor code to meet PEP8

+++ main.py
@@ -1,10 +1,7 @@
 import os
-import hashlib
-import hmac
-import random
-from string import letters
 
 # My files
+import auth
 from validate import Validate
 from datastore import Users
 from datastore import Blogs
@@ -17,23 +14,6 @@ import webapp2
 template_dir = os.path.join(os.path.dirname(__file__), "templates")
 jinja_env = jinja2.Environment(loader=jinja2.FileSystemLoader(template_dir),
                                autoescape=True)
-
-secret = "lkasdjf$j89u_345n45e-jtgdf8^459u23asd"
-
-
-def make_secure_val(val):
-    return "%s|%s" % (val, hmac.new(secret, val).hexdigest())
-
-
-def check_secure_val(secure_val):
-    val = secure_val.split("|")[0]
-    if secure_val == make_secure_val(val):
-        return val
-
-
-# Make salt for password hashing
-def make_salt(length=5):
-    return ''.join(random.choice(letters) for x in xrange(length))
 
 
 class Handler(webapp2.RequestHandler):
@@ -48,7 +28,7 @@ class Handler(webapp2.RequestHandler):
         self.write(self.render_str(template, **kw))
 
     def set_secure_cookie(self, name, val):
-        cookie_val = str(make_secure_val(val))
+        cookie_val = str(auth.make_secure_val(val))
         name = str(name)
         self.response.headers.add_header(
             "Set-Cookie",
@@ -57,7 +37,7 @@ class Handler(webapp2.RequestHandler):
 
     def read_secure_cookie(self, name):
         cookie_val = self.request.cookies.get(name)
-        return cookie_val and check_secure_val(cookie_val)
+        return cookie_val and auth.check_secure_val(cookie_val)
 
     # Turns away unauthorized users
     def user_page(self, origin_page, alt_page, **kw):
@@ -87,20 +67,6 @@ class MainPage(Handler):
     def get(self):
         blogs = Blogs.all().order("-created")
         self.render("index.html", blogs=blogs, auth_user=self.user)
-
-
-# Make the password hash with salt
-def make_pw_hash(name, pw, salt=None):
-    if not salt:
-        salt = make_salt()
-    h = hashlib.sha256(name + pw + salt).hexdigest()
-    return '%s,%s' % (salt, h)
-
-
-# Validate the password
-def valid_pw(name, password, h):
-    salt = h.split(',')[0]
-    return h == make_pw_hash(name, password, salt)
 
 
 class SignUpPage(Handler):
@@ -133,7 +99,7 @@ class SignUpPage(Handler):
         if not Validate(email).email():
             params['error_email'] = "Invalid email address"
             have_error = True
-        elif email == Users.by_email(email):
+        elif email and email == Users.by_email(email):
             params['error_email'] = "This email address is already used"
             have_error = True
 
@@ -141,7 +107,7 @@ class SignUpPage(Handler):
             self.render("signup.html", **params)
         else:
             # Create password hash
-            pass_hash = make_pw_hash(username, password)
+            pass_hash = auth.make_pw_hash(username, password)
 
             # Create user in DB
             user = Users(username=username, password=pass_hash, email=email)
@@ -152,6 +118,43 @@ class SignUpPage(Handler):
 
             # Redirect user to welcome page
             self.redirect("/welcome")
+
+
+class LogIn(Handler):
+    def get(self):
+        self.anom_page("login.html", "/welcome")
+
+    def post(self):
+        have_error = False
+        username = self.request.get("username")
+        password = self.request.get("password")
+
+        params = dict(username=username)
+
+        if self.valid_cred(username, password):
+            # Create cookie
+            self.set_secure_cookie("username", username)
+
+            # Redirect user to welcome page
+            self.redirect("/welcome")
+        else:
+            have_error = True
+            params['error_msg'] = "Invalid username or password"
+            self.render("login.html", **params)
+
+    def valid_cred(self, username, password, pass_hash=""):
+        query = Users.all().filter("username", username).get()
+        if query is not None:
+            pass_hash = query.password
+        return username and password and auth.valid_pw(username,
+                                                       password,
+                                                       pass_hash)
+
+
+class LogOut(Handler):
+    def get(self):
+        self.logout()
+        self.redirect("/")
 
 
 class WelcomePage(Handler):
@@ -233,6 +236,9 @@ class EditPost(Handler):
 
 
 class DeletePost(Handler):
+    def get(self):
+        self.error(404)
+
     def post(self):
         blog_id = self.request.get("bid")
         if Blogs.verify_owner(blog_id, self.user):
@@ -243,54 +249,45 @@ class DeletePost(Handler):
             self.redirect("/welcome")
 
 
-class LogIn(Handler):
-    def get(self):
-        self.anom_page("login.html", "/welcome")
-
-    def post(self):
-        have_error = False
-        username = self.request.get("username")
-        password = self.request.get("password")
-
-        params = dict(username=username)
-
-        if self.valid_cred(username, password):
-            # Create cookie
-            self.set_secure_cookie("username", username)
-
-            # Redirect user to welcome page
-            self.redirect("/welcome")
-        else:
-            have_error = True
-            params['error_msg'] = "Invalid username or password"
-            self.render("login.html", **params)
-
-    def valid_cred(self, username, password, pass_hash=""):
-        query = Users.all().filter("username", username).get()
-        if query is not None:
-            pass_hash = query.password
-        return username and password and valid_pw(username,
-                                                  password,
-                                                  pass_hash)
-
-
-class LogOut(Handler):
-    def get(self):
-        self.logout()
-        self.redirect("/")
-
-
 class SinglePost(Handler):
     def get(self, blog_id):
         blog = Blogs.by_id(blog_id)
-        sentiment = Sentiment.all().filter("username = ", self.user).filter("blog_id =", int(blog_id)).get()
-        comments = Comments.all().filter("blog_id = ", int(blog_id)).order("created")
+        sentiment = Sentiment.by_owner(self.user, blog_id)
+        comments = Comments.by_blog_id(blog_id)
+
+        params = dict(blog=blog,
+                      sentiment=sentiment,
+                      comments=comments,
+                      auth_user=self.user)
 
         if not blog:
             self.error(404)
             return
 
-        self.render("singlepost.html", blog=blog, sentiment=sentiment, auth_user=self.user, comments=comments)
+        self.render("singlepost.html", **params)
+
+    def post(self, blog_id):
+        blog_id = int(self.request.get("bid"))
+        comment = self.request.get("comment")
+        blog = Blogs.by_id(blog_id)
+        comments = Comments.by_blog_id(blog_id)
+        sentiment = Sentiment.by_owner(self.user, blog_id)
+
+        params = dict(blog_id=blog_id,
+                      auth_user=self.user,
+                      blog=blog,
+                      comments=comments,
+                      sentiment=sentiment)
+
+        if not comment:
+            params['error_comment'] = "You didn't write any comment!"
+            self.render("singlepost.html", **params)
+        else:
+            comment = Comments(blog_id=blog_id,
+                               comment=comment,
+                               username=self.user)
+            comment.put()
+            self.redirect("/%s#Comments" % blog_id)
 
 
 class LikePost(Handler):
@@ -298,9 +295,11 @@ class LikePost(Handler):
         blog_id = int(self.request.get("bid"))
 
         # Check to see if the user has interacted with this post before.
-        q = Sentiment.all().filter("username = ", self.user).filter("blog_id = ", blog_id).get()
-        if q is None:
-            sentiment = Sentiment(username=self.user, blog_id=blog_id, sentiment=True)
+        sentiment = Sentiment.by_owner(self.user, blog_id)
+        if sentiment is None:
+            sentiment = Sentiment(username=self.user,
+                                  blog_id=blog_id,
+                                  sentiment=True)
             sentiment.put()
 
             # Update the Datastore
@@ -316,9 +315,11 @@ class DislikePost(Handler):
         blog_id = int(self.request.get("bid"))
 
         # Check to see if the user has interacted with this post before.
-        q = Sentiment.all().filter("username = ", self.user).filter("blog_id = ", blog_id).get()
-        if q is None:
-            sentiment = Sentiment(username=self.user, blog_id=blog_id, sentiment=False)
+        sentiment = Sentiment.by_owner(self.user, blog_id)
+        if sentiment is None:
+            sentiment = Sentiment(username=self.user,
+                                  blog_id=blog_id,
+                                  sentiment=False)
             sentiment.put()
 
             # Update the Datastore
@@ -329,37 +330,34 @@ class DislikePost(Handler):
         self.redirect("/%s" % blog_id)
 
 
-class PostComment(Handler):
-    def post(self):
-        blog_id = int(self.request.get("bid"))
-        comment = self.request.get("comment")
-
-        comment = Comments(blog_id=blog_id, comment=comment, username=self.user)
-        comment.put()
-
-        self.redirect("/%s" % blog_id)
-
-
 class EditComment(Handler):
+    def get(self):
+        self.error(404)
+
     def post(self):
         comment_id = int(self.request.get("cid"))
         comment_text = self.request.get("comment")
 
         comment = Comments.by_id(comment_id)
-        comment.comment = comment_text
         blog_id = comment.blog_id
-        comment.put()
+        if comment and comment.username == self.user:
+            comment.comment = comment_text
+            comment.put()
 
         self.redirect("/%s" % blog_id)
 
 
 class DeleteComment(Handler):
+    def get(self):
+        self.error(404)
+
     def post(self):
         comment_id = int(self.request.get("cid"))
 
         comment = Comments.by_id(comment_id)
         blog_id = comment.blog_id
-        comment.delete()
+        if comment and comment.username == self.user:
+            comment.delete()
 
         self.redirect("/%s" % blog_id)
 
@@ -374,7 +372,6 @@ app = webapp2.WSGIApplication([
     ('/delete-post', DeletePost),
     ('/like-post', LikePost),
     ('/dislike-post', DislikePost),
-    ('/post-comment', PostComment),
     ('/edit-comment', EditComment),
     ('/delete-comment', DeleteComment),
     ('/login', LogIn),
